Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

Works for me, this is fascinating!

     min-height: calc(100vh - 140px);

Next please define this in a class somewhere in the top level, then we can add that class to our divs.  Please do NOT replicate it.  It should appear once.  Some class name like dashboard-body.

What is the 140px fudge factor? is that for the header and footer?

Then I will merge this.

Patch-set: 3
Label: Code-Review=-1