Update patch set 5
Patch Set 5: Code-Review-1

The new slide-out drawer solution is cool, but it adds an undesired behavior - now the vertical scroll bar appears when it should not, and once it appears, it never goes away.  What a pain in the neck!!!!  Have you seen this?  Do you have any ideas how to fix?

Patch-set: 5
Label: Code-Review=-1