Upgrade E2 to version 20190626
E2 added a method to get the RAN identities, which answers
a list of names that can be used to query the RAN status
endpoint.  Extend FE and BE to use the new information.
Drop the mock behavior for tracking connection requests;
add a mock behavior for RAN names to support R1 testing.

Change-Id: I669f1e0537676c8b2ea09e5b53080396d91d9868
Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>