Update patch set 3
Patch Set 3:

Hmm, ok in re-reading the schema I see the entries for the app descriptor.

But I will want to keep the door open for getting a LAYOUT to feed to this library, so we can tune/tweak a screen to make it pretty.  For that reason I think the back-end will need to be extended with an endpoint that fetches the JSON schema and possibly includes a layout in the answer.

Patch-set: 3