Update patch set 6
Patch Set 6:

(1 comment)

I think the layout element should provide user-friendly titles for all fields in the ANR config. For example, maybe the section title "The Controls Schema" could be better.

Do we have config elements, or can we make something up, for the Admission Control and Dual Connectivity apps that we demonstrate in the test server?

The whole dialog scrolls.  Is that by design, or could you possibly put up a dialog with a title that remains in place, and an area with a scrolling window?

Please refresh my memory: the catalog screen has CONFIG elements, and the control screen has CONTROL elements, is that right?  So this schema-driven UI is only for the CONFIG elements on the catalog screen?  Do you think we'll be using this schema-driven UI for any CONTROL elements?

Patch-set: 6
Label: Code-Review=0