Update patch set 12
Patch Set 12:

Nicolas, after the dynamically generated dialog disappears not status message is shown. I missed this behavior when I reviewed your code. Would you please extend so that after a user hits the SUBMIT button on the dialog with config properties, a "toast" with success or failure is displayed?

Patch-set: 12