Update patch set 7
Patch Set 7:

Thanks for repairing so many nitty-gritty code issues.  I'm trying out the feature and have comments about appearance and behavior.

Would you consider adding the Policy Control card on the same row as the existing cards on the Main screen grid?  Having to scroll down to see the next row isn't great.  Probably need to shrink the amount of space between the cards so that it can fit.  I tried column-gap without any success tho :/

On the Policy Control screen:
1. I am not a fan of the much darker grey background on the policy instances. 
2. There's no icon, but clicking on the left of the row triggers the expand-collapse behavior.  Could a control like possibly be added at the left, the way there are up and down arrows at the left of titles in the dialog, to make it obvious?  I think using an icon at the right isn't the right place to trigger expand-collapse.
3. Maybe for create instance you use a + or other suggestion of adding something; the wrench isn't the best choice of icon here because you're not just making an adjustment.

On the Policy Instance dialog:
1. Instead of showing just an acronym like ANR I would prefer to see "Edit Policy Instance". Maybe text can shown below that like "Edit ANR neighbor cell relation policy" ?  And what is the extra line "ANR-1"? 
2. I find it really confusing that the submit and close buttons are not at the bottom. I know you're using the special generated UI here, but is there any way to emulate the usual Angular dialog and put a dividing line then the Submit/Close buttons at the bottom? Can you tell the generated UI to skip decorating itself with buttons?  With both Properties and Json collapsed there is no control on the dialog to close it :(
3. The view-JSON feature is a powerful debugging tool, but does it need to be so prominent, and does it need to show the schema?  I'm guessing regular users will not use it.

On the old RIC Control screen, the xApp Control for Admissions Control seems to be broken. Clicking that row changes the screen to my custom AC xApp Policy UI, but with a dialog "Http failure 404 not found".  I'm guessing this is breakage in the front end since you evolved the A1 mediator controller API.  Please check it.

HTH

Patch-set: 7
Label: Code-Review=0