Update patch set 3
Patch Set 3:

applied -DfailIfNoTests=false as suggested in the error and it worked.

I see blank chart under AC and MC xApp stats. May be we can improve it to show some mock chart when the kibana URL is not found and show a warning message that kibana not configured.

Patch-set: 3
Label: Verified=0