Update patch set 11
Patch Set 11:

Well, you have now changed the app to show a user-please-wait dialog until the list of RIC instances is available.  Does this prevent the user from reaching a screen before the app is fully configured with the list?  Does this eliminate the race condition?  I think it does, and maybe you no longer need to subscribe to the list of instances first.

Patch-set: 11
Label: Code-Review=0