Update patch set 19
Patch Set 19:

I think it's totally fine for a controller class like AppControlComponent to save a copy of the instance key in the subscription callback method.  Then that value can be passed to undeploy or anything else where it's needed.  I think this is the proper information hiding. Only the controllers know which instance should be fetched, the service should not assume.

Patch-set: 19