Update patch set 21
Patch Set 21: Code-Review+1

Looks good, thanks for all the whitespace fixes. But it appears to me that you might still be extending this to additional components. For example I don't yet see an update to the E2 mgr controller, only to the E2 mgr service. Please tell me if you are still going, if you want me to continue working on this next week, or how you want to handle, thanks.

Patch-set: 21
Label: Code-Review=+1