Update patch set 23
Patch Set 23:

Thanks for making the changes.  I extended your getPath method slightly but cannot upload my change set due to the settings in this Gerrit.  I'm trying to get the settings adjusted but because of the holiday week, maybe nobody is working.  If you are working I could email a patch set to you to apply.  I don't want to upload a new change set because that looks like I'm stealing your work. :/

https://jira.linuxfoundation.org/servicedesk/customer/portal/2/IT-18596

Patch-set: 23