Update patch set 10
Patch Set 10:

There's a major open question here. In the new version of A1, the front-end can only change the Admission Control policy by sending a policy instance with a reference to the policy type ID. It COULD query the back-end to discover policy types and let the user choose one to edit.  Do you want to go that direction?

Patch-set: 10
Label: Code-Review=0