Update patch set 5
Patch Set 5:

I would like more votes on this, but I am not in favor of waiting here.  With several developers in a busy project like this we deal with conflicts as a routine matter, it's not a blocker.

Patch-set: 5
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>