Update patch set 4
Patch Set 4:

Nicolas this is now largely complete. The mock backend is trivial, we will have to extend it once you start a front-end implementation to support demos and such. Please review.

I think we have lost Manoop to his E911 project, so if you're comfortable with this, go ahead with the merge, then propagate to RanCoDev.

Patch-set: 4